Chapter 10
When to panic and when not to panic, that is the question of course. I think itll take a little bit of trial and error when it comes to panicing, and using Result<T, E> definitly a lot to unpack there. but with the FWF app I think Im going to get about as much practice as possible implementing database queries and https requests