Chapter 10.1 and 10.2
no files or code for 10.1 as it was just generics and I have done lots with that in and out of class.
Traits on the other hand does have a file. I feel like Traits will either be a large headache or amazing when I get to implementing them myself