10.3 Lifetimes
Lifetimes are intresting, because in c++ we just use our intuition and do our best not to reference memory once its been deallocated. A good programmer might never have this issue, and or its just a bug fix if it does happen. With rust making it absolutely necissary to annotate (unless it falls under the elision rules) it feels very strict. I would like to think im a good enough programmar not to have this issue. I mean I did write an entire blackjack simulator in c++, so as of now I feel like this is more chaining than freeing, but Id love to be wrong