chapter 17
Learned how to build and run nightly rust. also upgraded my rust version. Maybe I just missed it but the book did not warn me that edition2024 was needed and so I spent a bit of time figuring out how to get cargo to see the trpl crate i imported